+++ src/components/SchoolPicker/useSchoolPicker.test.tsx
@@ -46,7 +46,7 @@ describe("useSchoolPicker", () => {
     jest.clearAllMocks();
     jest.resetModules();
   });
-  test("Schools should be returned if fetch succeeds", async () => {
+  test("Schools should be returned with homeschool option if fetch succeeds", async () => {
     const { result, rerender } = renderHook(useSchoolPicker);
     act(() => result.current.setSchoolPickerInputValue("wes"));
     mockUseSWR.mockImplementationOnce(() => ({
@@ -54,7 +54,13 @@ describe("useSchoolPicker", () => {
       error: null,
     }));
     rerender();
-    expect(result.current.schools).toEqual(data);
+    expect(result.current.schools).toEqual([
+      ...data,
+      {
+        name: "Homeschool",
+        urn: "homeschool",
+      },
+    ]);
   });
   test("Schools not returned if fetch succeeds but searchterm.length < 2", async () => {
     mockUseSWR.mockImplementationOnce(() => ({
@@ -62,6 +68,7 @@ describe("useSchoolPicker", () => {
       error: null,
     }));
     const { result } = renderHook(() => useSchoolPicker());
+
     expect(result.current.schools).toEqual([]);
   });
   test("should throw an error if failed to fetch school ", async () => {